tg_id_bugfix

+++ bot/bot/models/service.py
@@ -6,7 +6,7 @@ from bot.models.base import Base
 class Service(Base):
     __tablename__ = "service"
     id = Column(BigInteger, primary_key=True, index=True)
-    telegram_id = Column(Integer)
+    telegram_id = Column(BigInteger)
     service_name = Column(String)
     login = Column(String)
     password = Column(String)
